renamed Locations.java and put Poll Locations classes into it's own package
