+++ app/src/main/java/com/example/android/bigappsvotenyc/ElectedOfficials/ElectedOfficialsActivity.java
@@ -1,4 +1,4 @@
-package com.example.android.bigappsvotenyc.ElectedRepresentatives;
+package com.example.android.bigappsvotenyc.ElectedOfficials;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -7,9 +7,9 @@ import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 
-import com.example.android.bigappsvotenyc.ElectedRepresentatives.controller.RepOfficialAdapter;
-import com.example.android.bigappsvotenyc.ElectedRepresentatives.model.RepResponse;
-import com.example.android.bigappsvotenyc.ElectedRepresentatives.service.RepService;
+import com.example.android.bigappsvotenyc.ElectedOfficials.controller.RepOfficialAdapter;
+import com.example.android.bigappsvotenyc.ElectedOfficials.model.RepResponse;
+import com.example.android.bigappsvotenyc.ElectedOfficials.service.RepService;
 import com.example.android.bigappsvotenyc.R;
 
 import retrofit2.Call;
@@ -22,9 +22,9 @@ import retrofit2.converter.gson.GsonConverterFactory;
  * Created by catwong on 1/29/17.
  */
 
-public class ElectedRepsActivity extends AppCompatActivity {
+public class ElectedOfficialsActivity extends AppCompatActivity {
 
-    private static final String TAG = ElectedRepsActivity.class.getSimpleName();
+    private static final String TAG = ElectedOfficialsActivity.class.getSimpleName();
     private static final String BASE_URL = "https://www.googleapis.com/civicinfo/v2/";
     private RecyclerView recyclerView;
     private RepOfficialAdapter officialAdapter;
