+++ app/src/main/java/com/example/android/bigappsvotenyc/ElectedOfficials/model/Channel.java
@@ -1,4 +1,4 @@
-package com.example.android.bigappsvotenyc.ElectedRepresentatives.model;
+package com.example.android.bigappsvotenyc.ElectedOfficials.model;
 
 /**
  * Created by catwong on 1/29/17.
