+++ app/src/main/java/com/example/android/bigappsvotenyc/MainFragment.java
@@ -9,8 +9,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 
-import com.example.android.bigappsvotenyc.ElectedRepresentatives.ElectedRepsActivity;
-import com.example.android.bigappsvotenyc.PollingLocations.Locations;
+import com.example.android.bigappsvotenyc.ElectedOfficials.ElectedOfficialsActivity;
+import com.example.android.bigappsvotenyc.PollingLocations.PollLocationsActivity;
 
 /**
  * Created by catwong on 1/29/17.
@@ -51,7 +51,7 @@ public class MainFragment extends Fragment {
     }
 
     private void gotoPollLocations() {
-        Intent intent = new Intent(getActivity(), Locations.class);
+        Intent intent = new Intent(getActivity(), PollLocationsActivity.class);
         MainFragment.this.startActivity(intent);
     }
 
@@ -66,7 +66,7 @@ public class MainFragment extends Fragment {
     }
 
     private void representativesByInfo() {
-        Intent intent = new Intent(getActivity(), ElectedRepsActivity.class);
+        Intent intent = new Intent(getActivity(), ElectedOfficialsActivity.class);
         MainFragment.this.startActivity(intent);
     }
