--- app/src/main/java/com/example/android/bigappsvotenyc/NextElection.java
@@ -1,5 +1,0 @@
-package com.example.android.bigappsvotenyc;
-
-public class NextElection {
-
-}
